reclassifying types of fields to match exercise description

+++ src/main/java/ca/abdullahr2/springstore/entities/Category.java
@@ -14,8 +14,10 @@ public class Category {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private int id;
+    private Byte id;
 
     @Column(name = "name")
     private String name;
+
+
 }
